Add type import to indicate importing only type
