+++ types/children.d.ts
@@ -1,4 +1,4 @@
-import { ReactNode } from 'react';
+import type { ReactNode } from 'react';
 
 export type Children = {
   children: ReactNode | React.ReactNode[];
